+++ packages/devtools_app/lib/src/screens/memory/area_controls/controls_widgets.dart
@@ -21,6 +21,7 @@ import '../memory_events_pane.dart' as events;
 import '../memory_vm_chart.dart' as vm;
 import '../primitives/painting.dart';
 import 'constants.dart';
+import 'legend.dart';
 import 'memory_config.dart';
 
 class ChartControls extends StatefulWidget {
@@ -292,23 +293,156 @@ class CommonControls extends StatefulWidget {
   const CommonControls(
       {Key? key,
       required this.isAndroidCollection,
-      required this.isAdvancedSettingsEnabled})
+      required this.isAdvancedSettingsEnabled,
+      required this.chartControllers})
       : super(key: key);
 
   final bool isAndroidCollection;
   final bool isAdvancedSettingsEnabled;
+  final ChartControllers chartControllers;
 
   @override
   State<CommonControls> createState() => _CommonControlsState();
 }
 
 class _CommonControlsState extends State<CommonControls>
-    with MemoryControllerMixin {
+    with MemoryControllerMixin, AutoDisposeMixin {
+  OverlayEntry? _legendOverlayEntry;
+
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
 
     initMemoryController();
+
+    addAutoDisposeListener(memoryController.legendVisibleNotifier, () {
+      setState(() {
+        if (memoryController.isLegendVisible) {
+          ga.select(
+            analytics_constants.memory,
+            analytics_constants.memoryLegend,
+          );
+
+          _showLegend(context);
+        } else {
+          _hideLegend();
+        }
+      });
+    });
+
+    addAutoDisposeListener(memoryController.androidChartVisibleNotifier, () {
+      setState(() {
+        if (memoryController.androidChartVisibleNotifier.value) {
+          ga.select(
+            analytics_constants.memory,
+            analytics_constants.androidChart,
+          );
+        }
+        if (memoryController.isLegendVisible) {
+          // Recompute the legend with the new traces now visible.
+          _hideLegend();
+          _showLegend(context);
+        }
+      });
+    });
+  }
+
+  void _showLegend(BuildContext context) {
+    final box = legendKey.currentContext!.findRenderObject() as RenderBox;
+
+    final colorScheme = Theme.of(context).colorScheme;
+    final legendHeading = colorScheme.hoverTextStyle;
+
+    // Global position.
+    final position = box.localToGlobal(Offset.zero);
+
+    final legendRows = <Widget>[];
+
+    final events = eventLegend(colorScheme.isLight);
+    legendRows.add(
+      Container(
+        padding: legendTitlePadding,
+        child: Text('Events Legend', style: legendHeading),
+      ),
+    );
+
+    var iterator = events.entries.iterator;
+    while (iterator.moveNext()) {
+      final leftEntry = iterator.current;
+      final rightEntry = iterator.moveNext() ? iterator.current : null;
+      legendRows.add(LegendRow(
+        entry1: leftEntry,
+        entry2: rightEntry,
+        chartControllers: widget.chartControllers,
+      ));
+    }
+
+    final vms = vmLegend(widget.chartControllers.vm);
+    legendRows.add(
+      Container(
+        padding: legendTitlePadding,
+        child: Text('Memory Legend', style: legendHeading),
+      ),
+    );
+
+    iterator = vms.entries.iterator;
+    while (iterator.moveNext()) {
+      final legendEntry = iterator.current;
+      legendRows.add(LegendRow(
+        entry1: legendEntry,
+        chartControllers: widget.chartControllers,
+      ));
+    }
+
+    if (memoryController.isAndroidChartVisible) {
+      final androids = androidLegend(widget.chartControllers.android);
+      legendRows.add(
+        Container(
+          padding: legendTitlePadding,
+          child: Text('Android Legend', style: legendHeading),
+        ),
+      );
+
+      iterator = androids.entries.iterator;
+      while (iterator.moveNext()) {
+        final legendEntry = iterator.current;
+        legendRows.add(LegendRow(
+          entry1: legendEntry,
+          chartControllers: widget.chartControllers,
+        ));
+      }
+
+      final OverlayState overlayState = Overlay.of(context)!;
+      _legendOverlayEntry ??= OverlayEntry(
+        builder: (context) => Positioned(
+          top: position.dy + box.size.height + legendYOffset,
+          left: position.dx - legendWidth + box.size.width - legendXOffset,
+          height: memoryController.isAndroidChartVisible
+              ? legendHeight2Charts
+              : legendHeight1Chart,
+          child: Container(
+            padding: const EdgeInsets.fromLTRB(0, 5, 5, 8),
+            decoration: BoxDecoration(
+              color: colorScheme.defaultBackgroundColor,
+              border: Border.all(color: Colors.yellow),
+              borderRadius: BorderRadius.circular(10.0),
+            ),
+            width: legendWidth,
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: legendRows,
+            ),
+          ),
+        ),
+      );
+
+      overlayState.insert(_legendOverlayEntry!);
+    }
+  }
+
+  void _hideLegend() {
+    _legendOverlayEntry?.remove();
+    _legendOverlayEntry = null;
   }
 
   @override
@@ -363,6 +497,23 @@ class _CommonControlsState extends State<CommonControls>
         ),
       ],
     );
+  }
+
+  void _openSettingsDialog() {
+    showDialog(
+      context: context,
+      builder: (context) => MemoryConfigurationsDialog(memoryController),
+    );
+  }
+
+  void _exportToFile() {
+    final outputPath = memoryController.memoryLog.exportMemory();
+    final notificationsState = Notifications.of(context);
+    if (notificationsState != null) {
+      notificationsState.push(
+        'Successfully exported file ${outputPath.last} to ${outputPath.first} directory',
+      );
+    }
   }
 
   Future<void> _gc() async {
